feat(calibration): Add configuration parsing and parameter management for robot calibration
